+++ sobolevsky.misha/S5/errorsANDinput.hpp
@@ -6,10 +6,8 @@
 namespace sobolevsky
 {
   void inputFromFile(std::istream &in, sobolevsky::AVLtree< long long, std::string, long long > &data);
-  void errorInvalidCommand(std::ostream &out);
   void errorEmpty(std::ostream &out);
   void errorOverflow(std::ostream &out);
-  void errorUnderflow(std::ostream &out);
 }
 
 #endif
