+++ sobolevsky.misha/common/queue.hpp
@@ -3,7 +3,7 @@
 
 #include <utility>
 #include <stdexcept>
-#include "list.hpp"
+#include <list.hpp>
 
 namespace sobolevsky
 {
