+++ sobolevsky.misha/common/stack.hpp
@@ -4,7 +4,7 @@
 #include <utility>
 #include <stdexcept>
 #include <memory>
-#include "list.hpp"
+#include <list.hpp>
 
 namespace sobolevsky
 {
